+++ NEWS.md
@@ -1,6 +1,7 @@
 # aqp 1.41 (2022-01-19)
  * `plotSPC()` gains new arguments for controlling horizon depth labels
  * `plotSPC()` gains horizon depth annotation collision fixes (https://github.com/ncss-tech/aqp/issues/240)
+ * `plotSPC()` minor bugfix for cases when `n != length(x)`
 
 # aqp 1.40 (2022-01-04)
  * CRAN release